+++ include/rtkZengForwardProjectionImageFilter.h
@@ -106,12 +106,12 @@ public:
   itkTypeMacro(ZengForwardProjectionImageFilter, ForwardProjectionImageFilter);
 
   /** Get / Set the sigma zero of the PSF. Default is 1.5417233052142099 */
-  itkGetMacro(SigmaZero, float);
-  itkSetMacro(SigmaZero, float);
+  itkGetMacro(SigmaZero, double);
+  itkSetMacro(SigmaZero, double);
 
   /** Get / Set the alpha of the PSF. Default is 0.016241189545787734 */
-  itkGetMacro(Alpha, float);
-  itkSetMacro(Alpha, float);
+  itkGetMacro(Alpha, double);
+  itkSetMacro(Alpha, double);
 
 protected:
   ZengForwardProjectionImageFilter();
@@ -153,8 +153,8 @@ private:
   void
   operator=(const Self &) = delete; // purposely not implemented
 
-  float      m_SigmaZero{ 1.5417233052142099 };
-  float      m_Alpha{ 0.016241189545787734 };
+  double     m_SigmaZero{ 1.5417233052142099 };
+  double     m_Alpha{ 0.016241189545787734 };
   VectorType m_VectorOrthogonalDetector{ 0. };
   PointType  m_centerVolume{ 0 };
 };
